requirements.txt
groq
python-dotenv
langgraph
langchain_community
langchain_groq
langchain_ollama